% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocking.R
\name{AdvancedBlock}
\alias{AdvancedBlock}
\title{AdvancedBlock}
\usage{
AdvancedBlock(dt, type)
}
\arguments{
\item{dt}{a datatable}

\item{type}{Binary, Categorical, Neither.}
}
\value{
a vector of index
}
\description{
AdvancedBlock
}
