% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocking.R
\name{auto_matching}
\alias{auto_matching}
\title{the ultimate function that combines the results of find_best_match_based_on_KS_test and find_best_match_character_features}
\usage{
auto_matching(
  dt1,
  dt2,
  sample_size_numeric = 500,
  sample_size_string = 500,
  characters_to_match = letters
)
}
\arguments{
\item{dt1}{a data table}

\item{dt2}{a data table}

\item{sample_size_numeric}{sample size to reduce runtime}

\item{sample_size_string}{sample size to reduce runtime}

\item{characters_to_match}{letters}
}
\value{
returns a numeric vector of length equal to length(dt1); c(3,0,1) means the first column in dt1 matched with the third column in dt2, the second column in dt1 couldn't be matched with any column in dt2, and the third column in dt1 is matched with the first column on dt2.
}
\description{
the ultimate function that combines the results of find_best_match_based_on_KS_test and find_best_match_character_features
}
