% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocking.R
\name{BaseBlock}
\alias{BaseBlock}
\title{BaseBlock}
\usage{
BaseBlock(dt, type = "All")
}
\arguments{
\item{dt}{a data.table}

\item{type}{Unmatchable, Time, String, Int, Double, ALL.}
}
\value{
a vector of index
}
\description{
BaseBlock
}
