% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocking.R
\name{CalculateCosine}
\alias{CalculateCosine}
\title{calculate the consine of vec1 and vec 2}
\usage{
CalculateCosine(vec1, vec2)
}
\arguments{
\item{vec1}{a numeric vector}

\item{vec2}{a numeric vector}
}
\value{
a numeric vector of lengh 1; the value is between 0 and 1
}
\description{
calculate the consine of vec1 and vec 2
}
