% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocking.R
\name{find_best_match_based_on_KS_test}
\alias{find_best_match_based_on_KS_test}
\title{This function returns a mapping: a vector of length = length(dt1)}
\usage{
find_best_match_based_on_KS_test(dt1, dt2, input_sample_size = 500, type)
}
\arguments{
\item{dt1}{a data table}

\item{dt2}{a data table}

\item{input_sample_size}{sample how many data points}

\item{type}{type = "Int" or "Double"}
}
\value{
a vector of matching result
}
\description{
This function returns a mapping: a vector of length = length(dt1)
}
