% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocking.R
\name{FindBestCharFeatures}
\alias{FindBestCharFeatures}
\title{This function returns the matching result, which is of the same length of dt1}
\usage{
FindBestCharFeatures(
  dt1,
  dt2,
  characters_to_match = letters,
  sample_size = 500
)
}
\arguments{
\item{dt1}{a data table}

\item{dt2}{a data table}

\item{characters_to_match}{characters to match}

\item{sample_size}{a numeric vector of length 1; use sample size to reduce computational time}
}
\value{
a numeric vector of the same length(dt1)
}
\description{
This function returns the matching result, which is of the same length of dt1
}
