% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocking.R
\name{FindDouble}
\alias{FindDouble}
\title{Helper function, Find Double features' index}
\usage{
FindDouble(dt)
}
\arguments{
\item{dt}{a data table}
}
\value{
a vector of index
}
\description{
Helper function, Find Double features' index
}
