% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocking.R
\name{FindInteger}
\alias{FindInteger}
\title{Helper function, Find Integer features' index}
\usage{
FindInteger(dt)
}
\arguments{
\item{dt}{a data table}
}
\value{
a vector of index
}
\description{
Helper function, Find Integer features' index
}
