% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocking.R
\name{FindString}
\alias{FindString}
\title{Helper function, Find String features' index}
\usage{
FindString(dt)
}
\arguments{
\item{dt}{a data table}
}
\value{
a vector of index
}
\description{
Helper function, Find String features' index
}
