% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocking.R
\name{FindTime}
\alias{FindTime}
\title{Helper function, Find time features' index}
\usage{
FindTime(dt)
}
\arguments{
\item{dt}{a datatable}
}
\value{
a vector of index
}
\description{
Helper function, Find time features' index
}
