% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocking.R
\name{FindUnmatchable}
\alias{FindUnmatchable}
\title{Helper Function, Find Unmatchable features' index}
\usage{
FindUnmatchable(dt)
}
\arguments{
\item{dt}{a datatable}
}
\value{
a vector of index
}
\description{
Helper Function, Find Unmatchable features' index
}
