% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocking.R
\name{get_occurrence_num}
\alias{get_occurrence_num}
\title{get_occurrence_num returns the number of occurrences of each character in characters_to_match for char_feature}
\usage{
get_occurrence_num(
  char_feature,
  characters_to_match = letters,
  input_sample_size = 500
)
}
\arguments{
\item{char_feature}{a character vector}

\item{characters_to_match}{a character vector}

\item{input_sample_size}{a numeric vector of length 1}
}
\value{
a numeric vector recording the number of occurrence of each character in characters_to_match
}
\description{
get_occurrence_num returns the number of occurrences of each character in characters_to_match for char_feature
}
