% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocking.R
\name{GetOccurrenceMatrixAsDF}
\alias{GetOccurrenceMatrixAsDF}
\title{returns a data frame containing the number of occurrences of each character in characters_to_match for every character columns in the input data table}
\usage{
GetOccurrenceMatrixAsDF(
  dt,
  characters_to_match = letters,
  input_sample_size = 500
)
}
\arguments{
\item{dt}{a data table}

\item{characters_to_match}{characters to match}

\item{input_sample_size}{a numeric vector of length 1; sample to reduce computational time}
}
\value{
data frame
}
\description{
returns a data frame containing the number of occurrences of each character in characters_to_match for every character columns in the input data table
}
