% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blocking.R
\name{KSFeatureComparison}
\alias{KSFeatureComparison}
\title{This function visualizes similarity comparison between numerical features.}
\usage{
KSFeatureComparison(feature1, feature2, input_sample_size = 100)
}
\arguments{
\item{feature1}{a column of data}

\item{feature2}{a column of data}

\item{input_sample_size}{how many data point to include}
}
\value{
a map
}
\description{
This function visualizes similarity comparison between numerical features.
}
